Gapps: gapps-config.txt
This module will create a file /persist/gapps-config.txt under
root and recovery.

That will permit a compatible Gapps to be flashed and to NOT install
PackageInstallerGoogle (and remove AOSP PackageInstaller)

This file can be override in multiple other places by the user so
the user of the ROM needs to be informed to update his own gapps-config.txt
file.